Rename example apps from netdash_* to example_*
