+++ apps_dev/example_devices_dummy/example_devices_dummy/api/urls.py
