+++ apps_dev/example_devices_dummy/setup.py
@@ -4,7 +4,7 @@ from setuptools import find_packages, setup
 os.chdir(os.path.normpath(os.path.join(os.path.abspath(__file__), os.pardir)))
 
 setup(
-    name='netdash-devices-netbox-api',
+    name='example-devices-dummy',
     version='0.0.0',
     packages=find_packages(),
     description='Example NetDash module',
