+++ apps_dev/example_devices_netbox_api/example_devices_netbox_api/api/views.py
@@ -7,7 +7,7 @@ from rest_framework.response import Response
 from rest_framework.reverse import reverse
 from rest_framework import viewsets
 
-from netdash_devices_netbox_api.utils import get_device, get_devices
+from example_devices_netbox_api.utils import get_device, get_devices
 
 
 class DeviceViewSet(viewsets.ViewSet):
@@ -17,7 +17,7 @@ class DeviceViewSet(viewsets.ViewSet):
 
     def __init__(self, *args, **kwargs):
         if not settings.NETBOX_API_URL:
-            raise ImproperlyConfigured('NETBOX_API_URL to use netdash_devices_netbox_api')
+            raise ImproperlyConfigured('NETBOX_API_URL to use example_devices_netbox_api')
         self.URL = settings.NETBOX_API_URL.lstrip('/')
         super().__init__(*args, **kwargs)
