+++ apps_dev/example_devices_netbox_api/example_devices_netbox_api/templates/device-list.html
