+++ apps_dev/example_devices_netbox_api/example_devices_netbox_api/views.py
@@ -4,7 +4,7 @@ from django.conf import settings
 from django.core.exceptions import ImproperlyConfigured
 from django.views.generic import TemplateView
 
-from netdash_devices_netbox_api.utils import get_device, get_devices
+from example_devices_netbox_api.utils import get_device, get_devices
 
 
 class DeviceList(TemplateView):
@@ -12,7 +12,7 @@ class DeviceList(TemplateView):
 
     def __init__(self, *args, **kwargs):
         if not settings.NETBOX_API_URL:
-            raise ImproperlyConfigured('NETBOX_API_URL to use netdash_devices_netbox_api')
+            raise ImproperlyConfigured('NETBOX_API_URL to use example_devices_netbox_api')
         self.URL = settings.NETBOX_API_URL.lstrip('/')
         super().__init__(*args, **kwargs)
