+++ apps_dev/example_devices_snmp/example_devices_snmp/api/urls.py
