+++ apps_dev/example_devices_snmp/example_devices_snmp/api/views.py
@@ -1,5 +1,5 @@
-from netdash_devices_snmp.models import SnmpDevice
-from netdash_devices_snmp.api.serializers import SnmpDeviceSerializer
+from example_devices_snmp.models import SnmpDevice
+from example_devices_snmp.api.serializers import SnmpDeviceSerializer
 
 from rest_framework import viewsets
