--- apps_dev/netdash_devices_netbox_api/MANIFEST.in
@@ -1,2 +1,0 @@
-recursive-include netdash_devices_netbox_api/static *
-recursive-include netdash_devices_netbox_api/templates *
